refactor: drop unused useArtifactAutoOpen hook, fold shouldAutoOpen into watcher
The hook had no callers: the SDK's artifact renderers go through the
registry path (covered by the watcher), and the chat-shell integration
that might need a non-registry path is deferred. Archived for later
reintroduction if needed. Its shouldAutoOpen predicate moves into
artifactAutoOpenWatcher.ts; the decision-matrix test folds into the
watcher suite (still 128 tests).
