+++ packages/react-headless/src/store/__tests__/artifactAutoOpenWatcher.test.ts
@@ -1,7 +1,20 @@
 import { describe, expect, it } from "vitest";
-import { evaluateRegisteredArtifacts } from "../artifactAutoOpenWatcher";
+import { evaluateRegisteredArtifacts, shouldAutoOpen } from "../artifactAutoOpenWatcher";
 import { createDetailedViewStore } from "../createDetailedViewStore";
 import type { ArtifactEntry } from "../threadContextTypes";
+
+describe("shouldAutoOpen", () => {
+  it.each([
+    ["open-on-mount", true, true],
+    ["open-on-mount", false, true],
+    ["auto-open", true, true],
+    ["auto-open", false, false],
+    ["overview", true, false],
+    ["overview", false, false],
+  ] as const)("mode %s, streaming %s → %s", (mode, isStreaming, expected) => {
+    expect(shouldAutoOpen(mode, isStreaming)).toBe(expected);
+  });
+});
 
 const entry = (id: string, version = 1): ArtifactEntry => ({
   id,
